+++ src/app/components/index.ts
@@ -1,0 +1,3 @@
+export * from './header/header.component';
+export * from './order-card/order-card.component';
+export * from './previous-order-modal/previous-order-modal.component';
