+++ src/app/guards/index.ts
@@ -1,0 +1,2 @@
+export * from './anonym.guard';
+export * from './auth.guard';
