+++ src/app/models/index.ts
@@ -1,0 +1,3 @@
+export * from './forms.model';
+export * from './order.model';
+export * from './user.model';
