+++ src/app/pages/index.ts
@@ -1,0 +1,4 @@
+export * from './all-orders/all-orders.component';
+export * from './login/login.component';
+export * from './order/order.component';
+export * from './orders-summary/orders-summary.component';
