+++ src/app/services/index.ts
@@ -1,0 +1,3 @@
+export * from './auth.service';
+export * from './order.service';
+export * from './users.service';
